% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_correction.R
\name{correct_outliers}
\alias{correct_outliers}
\title{Correct and optionally redistribute outliers in an \code{epi_tibble} object.}
\usage{
correct_outliers(
  x,
  var,
  outliers_col = "outliers",
  detection_method = NULL,
  distribute_strategy = c("none"),
  distribute_start = min(x$time_value),
  distribute_end = max(x$time_value),
  outliers = NULL,
  new_col_name = paste0(as.character(substitute(var)), "_corrected")
)
}
\arguments{
\item{x}{The \code{epi_tibble} object under consideration.}

\item{var}{The variable in \code{x} in which to correct outliers.}

\item{outliers_col}{The name of a variable in \code{x} containing information
about outliers (detection thresholds and replacement values), in the format
as returned by \code{detect_outliers}. The default value is "outliers".}

\item{distribute_strategy}{A string specifying how the difference between the
original value and the corrected value should be redistributed to other
times. This may be \code{"none"} to do no redistribution, \code{"prop"} to
redistribute proportionally to existing values (a larger amount of the
difference is distributed to dates where the signal has larger values),
\code{"zeros"} to redistribute to dates where the signal has a value of 0, or
\code{"equal"} to redistribute approximately equally across a range of dates.}

\item{distribute_start}{the earliest index value to which we may
redistribute a difference between the initial value and corrected value.
If this column is not provided or the value is missing in a given row, a
default of the earliest available index value will be used.}

\item{distribute_end}{the last index value to which we may
redistribute a difference between the initial value and corrected value.
If this column is not provided or the value is missing in a given row, a
default of the \code{time_value} of the outlier minus one time step will be
used.}

\item{outliers}{A tibble specifying outliers and corrections to make; if
this argument is provided, the \code{outliers_col}, \code{detection_method},
\code{distribute_strategy}, \code{distribute_start}, and \code{distribute_end} arguments
are all ignored.
This argument allows for a more flexible per-outlier strategy for outlier
correction and redistribution by supplying the following columns:
\itemize{
\item \verb{<key name>}: a column for each column in the \code{epi_signal} key, specifying
values identifying the series to adjust
\item \code{time_value}: the date or time of the outlier to be corrected
\item \code{replacement} <num>: the corrected value that will replace the outlier
\item \code{distribute_strategy} <str>: optional, as documented above
\item \code{distribute_start} <date>: optional, as documented above
\item \code{distribute_end} <date>: optional, as documented above
}}

\item{new_col_name}{String indicating the name of the new column that will
contain the results of outlier correction. Default is the value of the
supplied \code{var} argument concatenated with \verb{_corrected}. Note that
setting \code{new_col_name} equal to an existing column name will overwrite this
column.}
}
\value{
a new \code{epi_tibble} with corrected outliers
}
\description{
Outliers are replaced with a corrected value. Optionally, the difference
between the initial outlying value and the corrected value may be
redistributed to other times.
}
