% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_correction.R
\name{correct_outliers_one_grp}
\alias{correct_outliers_one_grp}
\title{Function to do outlier correction for one group.}
\usage{
correct_outliers_one_grp(
  .data_group,
  var,
  outliers,
  new_col_name,
  grouping_vars,
  ...
)
}
\arguments{
\item{.data_group}{epi_tibble for one group. should represent one combination
of key factors determining a reporting unit, e.g., one geo_type or one
combination of geo_type and age group}

\item{var}{The variable in \code{x} in which to look for outliers.}

\item{new_col_name}{String indicating the name of the new column that will
contain the results of outlier detection. Default is "outliers"; note that
setting \code{new_col_name} equal to an existing column name will overwrite this
column.}

\item{grouping_vars}{variables that defined groups in the original input to
correct_outliers. These will be dropped from the return value for this
function}
}
\description{
This function is not intended to be called by package users.
}
