% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epitools.R
\name{inter.coeff}
\alias{inter.coeff}
\title{inter.coeff}
\usage{
inter.coeff(mod, gp)
}
\arguments{
\item{mod}{a model include interaction fiited by coxme}

\item{factor}{a string indicate the estimated interacted factor}
}
\description{
This function allows you to caculate coefficient and confidence interval of each stratum of interaction term.
}
\examples{
inter.coeff(model,"pm25")
}
