% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epitools.R
\name{inter.reri}
\alias{inter.reri}
\title{inter.reri}
\usage{
inter.reri(model, coeff, type = c("RERI", "APAB", "S"), conf.level = 0.95)
}
\arguments{
\item{model}{a model include interaction fiited by coxme}

\item{coeff}{a vector indicate the location of joint risk AB, seperate risk A and seperate risk B}

\item{type}{a string indicate the needed statistics ("RERI", "APAB" or "S")}

\item{conf.level}{a numeric indicate the level of confidence interval}
}
\description{
This function allows you to caculate RERI, APAB and S with confidence interval of interaction term.
}
\examples{
inter.reri(model1,c(2,3,4),"RERI",0.95)
}
