% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_correction.R
\name{slice_outliers}
\alias{slice_outliers}
\title{Keep rows identified as outliers by a specified detection method}
\usage{
slice_outliers(x, var, outliers_col, detection_method)
}
\description{
Keep rows identified as outliers by a specified detection method
}
