\name{wcgs}
\docType{data}
\alias{wcgs}
\title{Western Collaborative Group Study data}
\description{
  The Western Collaborative Group Study (WCGS), a prospective cohort
  studye, recruited middle-aged men (ages 39 to 59) who were employees
  of 10 California companies and collected data on 3154 individuals
  during the years 1960-1961. These subjects were primarily selected to
  study the relationship between behavior pattern and the risk of
  coronary hearth disease (CHD). A number of other risk factors were
  also measured to provide the best possible assessment of the CHD risk
  associated with behavior type. Additional variables collected include
  age, height, weight, systolic blood pressure, diastolic blood
  pressure, cholesterol, smoking, and corneal arcus.
}
\usage{
##data(wcgs)
}
\format{
  \itemize{
    \item{\code{id}}{
      Subject ID: 
    }  
    \item{\code{age0}}{
      Age: age in years 
    }
    \item{\code{height0}}{
      Height: height in inches
    }
    \item{\code{weight0}}{
      Weight: weight in pounds 
    }  
    \item{\code{sbp0}}{
      Systolic blood pressure: mm Hg
    }
    \item{\code{dbp0}}{
      Diastolic blood pressure: mm Hg
    }
    \item{\code{chol0}}{
      Cholesterol: mg/100 ml
    }
    \item{\code{behpat0}}{
      Behavior pattern:
    }  
    \item{\code{ncigs0}}{
      Smoking: Cigarettes/day
    }
    \item{\code{dibpat0}}{
      Dichotomous behavior pattern: 0 = Type B; 1 = Type A
    }
    \item{\code{chd69}}{
      Coronary heart disease event: 0 = none; 1 = yes
    }
    \item{\code{typechd}}{
      to be done
    }  
    \item{\code{time169}}{
      Observation (follow up) time: Days 
    }
    \item{\code{arcus0}}{
      Corneal arcus: 0 = none; 1 = yes
    }
  }
}
\source{UC Berkeley School of Public Health}
\references{
  pending
}
\keyword{datasets}
