\name{wnv}
\docType{data}
\alias{wnv}
\title{West Nile Virus human cases reported in California, USA, as of December 14, 2004}
\description{
Public Health Surveillance data
}
\usage{
##data(wnv)
}
\format{pending}
\source{California Department of Health Services}
\references{
  pending
}
\keyword{datasets}
